cache: current height use cache

+++ migrations/20220422034846_init.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS "indexer_raw"
 (
     "hash"      text  PRIMARY KEY,
     "prev_hash" text  NOT NULL,
-    "height"    numeric NOT NULL,
+    "height"    bigint NOT NULL,
     "raw"       jsonb NOT NULL
 );
